requirements_dev.txt
# pre-commit
pre-commit >= 3.5.0

# testing
pytest >= 7.4.3

# linting
ruff >= 0.1.3

# docs
mkdocs <= 1.5.3
mkdocs-material <= 9.4.6
mkdocstrings <= 0.23.0
mkdocstrings[python]>=0.18

# config
hydra-core >= 1.3.2
torch_tb_profiler >= 0.4.3